requirements.txt
  
grpcio==1.33.2
protobuf==3.14.0
grpcio-tools==1.33.2
sklearn
